separate PostsRepository into its interface and implementation
