+++ FooS/app/src/main/java/com/example/foos/data/domain/fetcher/post/FetchPostByPostIdUseCase.kt
@@ -5,6 +5,7 @@ import com.example.foos.data.model.database.DatabaseReaction
 import com.example.foos.data.model.database.DatabaseUser
 import com.example.foos.data.model.Post
 import com.example.foos.data.repository.PostsRepository
+import com.example.foos.data.repository.PostsRepositoryImpl
 import com.example.foos.data.repository.ReactionsRepository
 import com.example.foos.data.repository.UsersRepository
 import kotlinx.coroutines.Job
