+++ FooS/app/src/main/java/com/example/foos/data/domain/fetcher/post/FetchPostsByLocationBoundsUseCase.kt
@@ -4,6 +4,7 @@ import com.example.foos.data.model.database.DatabaseReaction
 import com.example.foos.data.model.database.DatabaseUser
 import com.example.foos.data.model.Post
 import com.example.foos.data.repository.PostsRepository
+import com.example.foos.data.repository.PostsRepositoryImpl
 import com.example.foos.data.repository.ReactionsRepository
 import com.example.foos.data.repository.UsersRepository
 import com.google.android.gms.maps.model.LatLngBounds
