+++ FooS/app/src/main/java/com/example/foos/data/domain/fetcher/post/FetchPostsWithMediaByUserIdUseCase.kt
@@ -5,6 +5,7 @@ import com.example.foos.data.model.database.DatabasePost
 import com.example.foos.data.model.database.DatabaseReaction
 import com.example.foos.data.model.database.DatabaseUser
 import com.example.foos.data.repository.PostsRepository
+import com.example.foos.data.repository.PostsRepositoryImpl
 import com.example.foos.data.repository.ReactionsRepository
 import com.example.foos.data.repository.UsersRepository
 import kotlinx.coroutines.Job
@@ -28,7 +29,7 @@ class FetchPostsWithMediaByUserIdUseCase @Inject constructor(
 
         coroutineScope {
             jobs.add(async { dbUser = usersRepository.fetchByUserId(userId) })
-            jobs.add(async { dbPosts = postsRepository.fetchWithMediaByUserId(userId, start, end) })
+            jobs.add(async { dbPosts = postsRepository.fetchWithMediaByUserId(userId, start, end, PostsRepositoryImpl.DEFAULT_LOAD_LIMIT) })
         }
         jobs.joinAll()
         jobs.clear()
